build_system_setup.sh: Build all the libs in Debug mode.
The androiddeployqt tries to build a DEBUG APK by default. Building all
the libs in RELEASE mode and the APK in DEBUG mode results in a non-working
application bundle. Building everything in debug mode works.
Building libs+APK in release mode should be the way to go, but it requires
a keystore file, alias and password in order to create a valid APK.
Trying to run androiddeployqt without the above mentioned keys, will result
in an uninstallable APK.
Link to Qt docs: https://doc.qt.io/qt-5/deployment-android.html .

Signed-off-by: Alexandra Trifan <Alexandra.Trifan@analog.com>